Add files via upload
This is a small project I created during a workshop designed for beginners. The primary goal of this project was to apply the fundamental concepts I learned, such as basic HTML, CSS, PHP, and MySQL integration. It’s a simple registration form that allows users to input their details, which are then stored in a database.

Throughout the workshop, I focused on understanding how to structure a web page, style it using CSS, and manage form data with PHP. Although this is a basic project, it represents an important step in my learning journey, helping me gain hands-on experience with web development.